just from pure curiosity
i wondered what is the most likely sum of three dices and get the answer on the internet but about four and five dices there were nothing. So I figured it out myself. 
I will think of a better way to add the dices. Because now it is not ideal.